Program project collaboration through publication and citation

copublication_data() returns a tibble enumerating all pmid acknowledging more than one project.

copublication() summarizes collaboration between projects at the time of publication.

cocitation_data() returns a tibble enumerating all pmid cited by other program pmid.

cocitation() summarizes collaborations between projects through citation.

Usage

copublication_data(tbl, exclude = NULL)

copublication(tbl)

cocitation_data(tbl)

cocitation(tbl)

Arguments

tbl

a tibble with column full_foa containing the Funding Opportunity Announcements defining the projects of interest.

exclude

character() pmid to exclude from summary.

Value

copublication_data() returns a tibble with the following columns.

core_project_num.x, core_project_num.y: character() pairwise project collaboration.

n: intger() number of collaborations between projects.

copublication() returns a tibble with columns

core_project_num: character() project numbers involved in collaboration

n: integer() total number of publications from the project.

citn: integer() total number of citations to project publications.

rcr: numeric() total relative citation rate of project publications.

collab: integer() number of distinct collaborators, across all collaborative publications.

n_collab: integer() number of collaborative publications.

citn_collab: numeric() total 'citation_count' for collaborative publications.

rcr_collab: numeric() total 'relative citation ratio' for collaborative publications.

cocitation_data() returns a tibble with the following columns.

pmid: integer() pmid of original publication.

cited_by: integer() pmid of publication citing original publication.

core_project_num: character() project number of project acknowledged in original publication.

cited_by_core_project_num: character() project number of project citing the original publication.

cocitation() returns a tibble with columns

project: character() project number associated with publications.

n: integer() total number of publications from the project.

citn: integer() total number of citations to project publications.

rcr: numeric() total relative citation rate of project publications.

n_self_citn: integer() number of citations to other publications of the same project.

n_collab_citn: integer() number of publications cited by other projects.

n_collab: integer() total number of projects citing this project.

rcr_citn: numeric() total 'relative citation ratio' of collaborative cited publications.

Examples

foas <- tribble(
    ~full_foa,       ~description,
    "RFA-CA-19-039", "Early-Stage Development of ...",
    "RFA-CA-19-038", "Development of Innovative ...",
    "RFA-CA-19-040", "Advanced Development of ...",
    "RFA-CA-19-041", "Sustained Support of ..."
)

copublication_data(foas)
#> # A tibble: 0 × 3
#> # ℹ 3 variables: core_project_num.x <chr>, core_project_num.y <chr>, n <int>

copublication(foas)
#> # A tibble: 16 × 8
#>    core_project_num     n  citn    rcr collab n_collab citn_collab rcr_collab
#>    <chr>            <int> <dbl>  <dbl>  <int>    <int>       <dbl>      <dbl>
#>  1 R21CA248118          5    51   9.68     NA       NA          NA         NA
#>  2 R21CA248122          8    55   4.36     NA       NA          NA         NA
#>  3 R21CA253408          8    27   5.02     NA       NA          NA         NA
#>  4 R21CA253498         10    57  16.6      NA       NA          NA         NA
#>  5 U01CA247760          9    29   5.94     NA       NA          NA         NA
#>  6 U01CA248226         56   636 128.       NA       NA          NA         NA
#>  7 U01CA248235          5   192  34.4      NA       NA          NA         NA
#>  8 U01CA253403         16   149  21.6      NA       NA          NA         NA
#>  9 U01CA253511          5     2   0.56     NA       NA          NA         NA
#> 10 U24CA248010          3    38   3.67     NA       NA          NA         NA
#> 11 U24CA248265         15   284  41.9      NA       NA          NA         NA
#> 12 U24CA248453          4    24   6.34     NA       NA          NA         NA
#> 13 U24CA248454          8    63  22.7      NA       NA          NA         NA
#> 14 U24CA248457         32   713  51.8      NA       NA          NA         NA
#> 15 U24CA253531          4    34  12.1      NA       NA          NA         NA
#> 16 U24CA253539          1     0   0        NA       NA          NA         NA

cocite_data <- cocitation_data(foas)
cocite_data
#> # A tibble: 85 × 4
#>        pmid cited_by core_project_num cited_by_core_project_num
#>       <int>    <dbl> <chr>            <chr>                    
#>  1 29985391 31913126 U24CA248457      U24CA248457              
#>  2 29985391 30773340 U24CA248457      U24CA248457              
#>  3 30563849 35977718 U24CA248457      U24CA248457              
#>  4 30723579 34641956 U24CA248457      U24CA248457              
#>  5 30723579 32530570 U24CA248457      U24CA248457              
#>  6 30773340 31913126 U24CA248457      U24CA248457              
#>  7 30773340 31171722 U24CA248457      U24CA248457              
#>  8 30773340 34376580 U24CA248457      U24CA248457              
#>  9 31913126 34376580 U24CA248457      U24CA248457              
#> 10 31932491 34376580 U24CA248457      U24CA248457              
#> # ℹ 75 more rows

## pmid / core_project_num (and cited_by / cited_by_core_project_num)
## reflects the fact that a single pmid may acknowledge several
## projects
cocite_data |>
    distinct(pmid, core_project_num) |>
    count(pmid, sort = TRUE)
#> # A tibble: 55 × 2
#>        pmid     n
#>       <int> <int>
#>  1 29985391     1
#>  2 30563849     1
#>  3 30723579     1
#>  4 30773340     1
#>  5 31913126     1
#>  6 31932491     1
#>  7 32119649     1
#>  8 32142794     1
#>  9 32216127     1
#> 10 32679308     1
#> # ℹ 45 more rows

## cocitation 'edges' between projects, and their weights
cocite_data |>
    count(core_project_num, cited_by_core_project_num, sort = TRUE)
#> # A tibble: 13 × 3
#>    core_project_num cited_by_core_project_num     n
#>    <chr>            <chr>                     <int>
#>  1 U01CA248226      U01CA248226                  40
#>  2 U24CA248457      U24CA248457                  12
#>  3 R21CA248122      R21CA248122                  10
#>  4 R21CA248118      R21CA248118                   6
#>  5 U01CA253403      U01CA253403                   6
#>  6 R21CA253498      R21CA253498                   2
#>  7 U24CA248265      U24CA248265                   2
#>  8 U24CA253531      U24CA253531                   2
#>  9 R21CA253408      R21CA253408                   1
#> 10 U01CA248226      R21CA253498                   1
#> 11 U24CA248265      U24CA248453                   1
#> 12 U24CA248265      U24CA248457                   1
#> 13 U24CA248454      U24CA248454                   1

cocitation(foas)
#> # A tibble: 16 × 8
#>    core_project_num     n  citn    rcr n_self_citn n_collab_citn n_collab
#>    <chr>            <int> <dbl>  <dbl>       <int>         <int>    <int>
#>  1 R21CA248118          5    51   9.68           6            NA       NA
#>  2 R21CA248122          8    55   4.36          10            NA       NA
#>  3 R21CA253408          8    27   5.02           1            NA       NA
#>  4 R21CA253498         10    57  16.6            2            NA       NA
#>  5 U01CA247760          9    29   5.94          NA            NA       NA
#>  6 U01CA248226         56   636 128.            40             1        1
#>  7 U01CA248235          5   192  34.4           NA            NA       NA
#>  8 U01CA253403         16   149  21.6            6            NA       NA
#>  9 U01CA253511          5     2   0.56          NA            NA       NA
#> 10 U24CA248010          3    38   3.67          NA            NA       NA
#> 11 U24CA248265         15   284  41.9            2             2        2
#> 12 U24CA248453          4    24   6.34          NA            NA       NA
#> 13 U24CA248454          8    63  22.7            1            NA       NA
#> 14 U24CA248457         32   713  51.8           12            NA       NA
#> 15 U24CA253531          4    34  12.1            2            NA       NA
#> 16 U24CA253539          1     0   0             NA            NA       NA
#> # ℹ 1 more variable: rcr_citn <dbl>