Program-level queries and summaries

program_summary() finds all projects, publications, and citations funded by the Funding Opportunity Announcements (FOAs) in tbl, and summarizes by year the number of projects, total award ammount, publications, citation count, and relative citation ratio.

program_projects() returns a tibble of project numbers and standardized contact PI name and project title.

program_publications() retrieves basic information for all projects in the program.

Usage

program_summary(tbl, verbose = FALSE)

program_projects(tbl, by = c("foa", "project"), verbose = FALSE)

program_publications(tbl, verbose = FALSE)

Arguments

tbl

A tibble with column opportunity_number giving the FOAs that define the program.

verbose

logical(1) report additional detail about progress of reporter and icite queries.

by

character(1) indicating how program projects are summarized -- by foa so that projects funded by more than one FOA are reported for each FOA, or project so that the summary is by project number across FOA.

Value

program_summary() returns a tibble ordered by fiscal year summarizing project activity. Columns are

fiscal_year: integer() year of program.

program: integer() number of active projects funded by this program.

project: integer() number of projects funded by any program.

program_amount: integer() award amount to active projects from this program.

project_amount: integer() award amount across all programs.

publications: integer() number of publications.

citation_count: integer() citations to publications in year.

relative_citation_ratio: numeric() sum of relative citation ratios for all publications in year.

program_projects() returns a tibble summarizing projects funded under the FOAs. With by = "foa" (default), columns are:

opportunity_number: character() full FOA funding the project.

core_project_num: character() core project number.

project_start_date: date() start date of project

project_end_date: date() end date of project; maybe in the future

contact_pi_name: character() name of most-recent contact PI for FOA and core project number.

project_title: character() project title.

fiscal_year: integer() fiscal years of funding; may differ from project duration (end date - start date.

award_amount: integer() award amount through programs in the FOAs.

With by = "project", columns are the same but with project_start_date, project_end_date, fiscal_year, award_amount and most recent contact PI and project title summarized over all FOAs under which a project was funded.

program_publications() returns a tibble with columns:

opportunity_number: character() full FOA funding the project.

core_project_num: character() core project number.

pmid: integer() PubMed identifier.

year, title, authors, journal,doi: publication information.

citn: integer() number of publicatons citing this publication, from icite().

rcr: relative citation ratio, as defined by icite().

fcr: field citation rate, as defined by icite().

Details

program_summary() can return rows corresponding to years before and after the years the FOAs were active. Values before the FOA activity represent either a project funded through other programs, or a grantee assigning credit to the FOA for a citation published prior to the start of the FOA, or to a project funded by FOAs not included in the input tbl but with the same core project number as a funded project. Values after the end of the FOA represent publications that acknowledge the FOA after the termination of the grant.

program_projects() provides a single row for each project. It chooses as opportunity_number the most recent FOA under which the project was funded. It chooses as contact_pi_name and project_title the name of the contact PI and project title of the largest award in the most recent year of funding.

Examples

foas <- tribble(
    ~opportunity_number, ~description,
    "RFA-CA-19-039",     "Early-Stage Development of ...",
    "RFA-CA-19-038",     "Development of Innovative ...",
    "RFA-CA-19-040",     "Advanced Development of ...",
    "RFA-CA-19-041",     "Sustained Support of..."
)

program_summary(foas)
#> # A tibble: 6 × 8
#>   fiscal_year program project program_amount project_amount publication
#>         <int>   <int>   <int>          <int>          <int>       <int>
#> 1        2018      NA      NA             NA             NA           3
#> 2        2019      NA      NA             NA             NA           7
#> 3        2020      17      18        9688464        9870795          31
#> 4        2021      14      16        8376100        8532608          81
#> 5        2022      13      16        7924287        8427808          70
#> 6        2023       8       8        5809339        5809339          70
#> # ℹ 2 more variables: citation_count <int>, relative_citation_ratio <dbl>

program_projects(foas)
#> # A tibble: 17 × 9
#>    opportunity_number core_project_num project_start_date project_end_date
#>    <chr>              <chr>            <date>             <date>          
#>  1 RFA-CA-19-038      R21CA248118      2020-08-01         2022-11-17      
#>  2 RFA-CA-19-038      R21CA248122      2020-05-15         2023-04-30      
#>  3 RFA-CA-19-038      R21CA253408      2020-09-01         2024-08-31      
#>  4 RFA-CA-19-038      R21CA253498      2020-09-01         2024-08-31      
#>  5 RFA-CA-19-039      U01CA247760      2020-09-01         2024-05-31      
#>  6 RFA-CA-19-039      U01CA248226      2020-07-01         2024-06-30      
#>  7 RFA-CA-19-039      U01CA248235      2020-08-01         2024-07-31      
#>  8 RFA-CA-19-039      U01CA253403      2020-09-01         2024-08-31      
#>  9 RFA-CA-19-039      U01CA253511      2020-08-15         2023-07-31      
#> 10 RFA-CA-19-040      U24CA248010      2020-09-24         2025-08-31      
#> 11 RFA-CA-19-040      U24CA248265      2020-07-01         2025-06-30      
#> 12 RFA-CA-19-040      U24CA248453      2020-09-24         2025-08-31      
#> 13 RFA-CA-19-040      U24CA248454      2020-07-15         2025-06-30      
#> 14 RFA-CA-19-040      U24CA248455      2020-09-03         2025-08-31      
#> 15 RFA-CA-19-040      U24CA253531      2020-09-01         2025-08-31      
#> 16 RFA-CA-19-040      U24CA253539      2020-09-10         2025-08-31      
#> 17 RFA-CA-19-041      U24CA248457      2020-06-01         2025-05-31      
#> # ℹ 5 more variables: contact_pi_name <chr>, project_title <chr>,
#> #   fiscal_year <int>, award_amount <int>, project_amount <int>

pubs <- program_publications(foas)

## note that individual publications can be represented by more
## than one project
pubs
#> # A tibble: 262 × 11
#>    opportunity_number core_project_num    pmid  year title authors journal doi  
#>    <chr>              <chr>              <int> <int> <chr> <chr>   <chr>   <chr>
#>  1 RFA-CA-19-040      U24CA253539       3.64e7  2023 PDCM… Zinaid… Nuclei… 10.1…
#>  2 RFA-CA-19-040      U24CA253531       3.68e7  2023 Eval… Dariya… Tomogr… 10.3…
#>  3 RFA-CA-19-040      U24CA253531       3.70e7  2023 Anim… Donna … Tomogr… 10.3…
#>  4 RFA-CA-19-040      U24CA253531       3.72e7  2023 Co-C… Stephe… Tomogr… 10.3…
#>  5 RFA-CA-19-040      U24CA253531       3.44e7  2021 Deep… Kaushi… Cancer… 10.3…
#>  6 RFA-CA-19-040      U24CA253531       3.43e7  2022 Co-c… Sudipt… Eur J … 10.1…
#>  7 RFA-CA-19-040      U24CA253531       3.71e7  2023 An O… Seth T… Tomogr… 10.3…
#>  8 RFA-CA-19-041      U24CA248457       3.13e7  2019 Metf… Xingyu… Cancer… 10.1…
#>  9 RFA-CA-19-041      U24CA248457       3.19e7  2020 STRI… Jong W… Elife   10.7…
#> 10 RFA-CA-19-041      U24CA248457       3.21e7  2020 TEAD… Jingti… J Inve… 10.1…
#> # ℹ 252 more rows
#> # ℹ 3 more variables: citation_count <int>, relative_citation_ratio <dbl>,
#> #   field_citation_rate <dbl>

## unique publications
unique_pubs <-
    pubs |>
    select(-c("opportunity_number", "core_project_num")) |>
    distinct()
unique_pubs
#> # A tibble: 262 × 9
#>        pmid  year title                     authors journal doi   citation_count
#>       <int> <int> <chr>                     <chr>   <chr>   <chr>          <int>
#>  1 36399494  2023 PDCM Finder: an open glo… Zinaid… Nuclei… 10.1…              3
#>  2 36828382  2023 Evaluation of Apparent D… Dariya… Tomogr… 10.3…              3
#>  3 36961012  2023 Animal Models and Their … Donna … Tomogr… 10.3…              2
#>  4 37218941  2023 Co-Clinical Imaging Meta… Stephe… Tomogr… 10.3…              2
#>  5 34359696  2021 Deep Learning Segmentati… Kaushi… Cancer… 10.3…             17
#>  6 34328530  2022 Co-clinical FDG-PET radi… Sudipt… Eur J … 10.1…             31
#>  7 37104131  2023 An Online Repository for… Seth T… Tomogr… 10.3…              2
#>  8 31292160  2019 Metformin Inhibits Progr… Xingyu… Cancer… 10.1…             27
#>  9 31913126  2020 STRIPAK directs PP2A act… Jong W… Elife   10.7…             32
#> 10 32142794  2020 TEAD1 and TEAD3 Play Red… Jingti… J Inve… 10.1…             12
#> # ℹ 252 more rows
#> # ℹ 2 more variables: relative_citation_ratio <dbl>, field_citation_rate <dbl>

## publications per project
pubs |>
    count(core_project_num, sort = TRUE) |>
    left_join(program_projects(foas, by = "project"), by = "core_project_num")
#> # A tibble: 16 × 10
#>    core_project_num     n opportunity_number project_start_date project_end_date
#>    <chr>            <int> <chr>              <date>             <date>          
#>  1 U01CA248226         74 RFA-CA-19-039      2020-07-01         2024-06-30      
#>  2 U24CA248457         35 RFA-CA-19-041      2020-06-01         2025-05-31      
#>  3 U01CA253403         24 RFA-CA-19-039      2020-09-01         2024-08-31      
#>  4 U24CA248265         18 RFA-CA-19-040      2020-07-01         2025-06-30      
#>  5 R21CA253498         16 RFA-CA-19-038      2020-09-01         2024-08-31      
#>  6 U24CA248454         15 RFA-CA-19-040      2020-07-15         2025-06-30      
#>  7 U24CA248453         14 RFA-CA-19-040      2020-09-24         2025-08-31      
#>  8 R21CA253408         12 RFA-CA-19-038      2020-09-01         2024-08-31      
#>  9 U01CA248235         12 RFA-CA-19-039      2020-08-01         2024-07-31      
#> 10 U01CA247760         10 RFA-CA-19-039      2020-09-01         2024-05-31      
#> 11 R21CA248122          9 RFA-CA-19-038      2020-05-15         2023-04-30      
#> 12 R21CA248118          6 RFA-CA-19-038      2020-08-01         2022-11-17      
#> 13 U24CA253531          6 RFA-CA-19-040      2020-09-01         2025-08-31      
#> 14 U01CA253511          5 RFA-CA-19-039      2020-08-15         2023-07-31      
#> 15 U24CA248010          5 RFA-CA-19-040      2020-09-24         2025-08-31      
#> 16 U24CA253539          1 RFA-CA-19-040      2020-09-10         2025-08-31      
#> # ℹ 5 more variables: contact_pi_name <chr>, project_title <chr>,
#> #   fiscal_year <int>, award_amount <int>, project_amount <int>

## collaborations; see `?copublication`, `?cocitation`
pubs |>
    count(pmid, sort = TRUE, name = "n_collab") |>
    left_join(unique_pubs, by = "pmid")
#> # A tibble: 262 × 10
#>        pmid n_collab  year title            authors journal doi   citation_count
#>       <int>    <int> <int> <chr>            <chr>   <chr>   <chr>          <int>
#>  1 29985391        1  2018 Assembly and ac… Daniel… Nat Co… 10.1…            101
#>  2 30563849        1  2019 A Curated Resou… Karste… Mol Ce… 10.1…            127
#>  3 30583461        1  2018 Comprehensive G… Natali… Cancer… 10.3…             20
#>  4 30723579        1  2018 APOBEC-related … Amélie… Oncoim… 10.1…             48
#>  5 30773340        1  2019 A Platform of S… Xiaodo… Cancer… 10.1…            134
#>  6 31171722        1  2019 Illuminating th… Victor… J Biol… 10.1…            109
#>  7 31292160        1  2019 Metformin Inhib… Xingyu… Cancer… 10.1…             27
#>  8 31395609        1  2019 Use of Natural … Guerga… Cancer… 10.1…             54
#>  9 31513970        1  2019 Cyclin E Overex… Servan… iScien… 10.1…              3
#> 10 31695841        1  2019 The multikinase… Sean M… Oncota… 10.1…              8
#> # ℹ 252 more rows
#> # ℹ 2 more variables: relative_citation_ratio <dbl>, field_citation_rate <dbl>